test: rename packages to mockebike
