+++ app/src/test/java/com/eddy/mockebike/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.eddy.nrf
+package com.eddy.mockebike
 
 import org.junit.Test
